# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "version": "1.0.0",
        "title": "Just Her Rideshare",
        "description": "Just Her Rideshare",
        "contact": {
            "name": "Smart Data Team"
        }
    },
    "paths": {
...
        "/driver/details": {
            "get": {
                "x-controller-name": "DriverController",
                "x-operation-name": "Details",
                "tags": ["DriverController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "Details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "parameters": [{
                    "name": "code",
                    "in": "query",
                    "schema": {
                        "type": "string"
                    }
                }, {
                    "name": "driver_id",
                    "in": "query",
                    "schema": {
                        "type": "string"
                    }
                }],
                "operationId": "DriverController.Details"
            }
        },
...
        "/trip/dispute": {
            "post": {
                "x-controller-name": "TripController",
                "x-operation-name": "Dispute",
                "tags": ["TripController"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "Dispute",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "code": {
                                        "type": "string"
                                    },
                                    "customer_id": {
                                        "type": "string"
                                    },
                                    "trip_id": {
                                        "type": "string"
                                    },
                                    "dispute_msg": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "operationId": "TripController.Dispute"
            }
        },
...
        "/common/maker-list": {
            "get": {
                "x-controller-name": "MakeModel",
                "x-operation-name": "MakerList",
                "tags": ["MakeModel"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "MakerList",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "parameters": [{
                    "name": "code",
                    "in": "query",
                    "schema": {
                        "type": "string"
                    }
                }],
                "operationId": "MakeModel.MakerList"
            }
        },
        "/common/model-list": {
            "get": {
                "x-controller-name": "MakeModel",
                "x-operation-name": "ModelList",
                "tags": ["MakeModel"],
                "security": [{
                    "jwt": []
                }],
                "responses": {
                    "200": {
                        "description": "ModelList",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {}
                                }
                            }
                        }
                    }
                },
                "parameters": [{
                    "name": "code",
                    "in": "query",
                    "schema": {
                        "type": "string"
                    }
                }, {
                    "name": "make_id",
                    "in": "query",
                    "schema": {
                        "type": "string"
                    }
                }],
                "operationId": "MakeModel.ModelList"
            }
        }
    },
...
    "servers": [{
        "url": "http://52.70.114.84:9133/api"
    }, {
        "url": "http://localhost:9133/api"
    }]
}